requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
torch
httpx
openai
python-dotenv
sqlalchemy
websockets
